+++ src/main/java/com/plivo/api/models/verify_session/AttemptDetail.java
@@ -5,12 +5,16 @@ public class AttemptDetail {
   private String attemptUuid;
   private String status;
   private String time;
+  private String brandName;
+  private String appHash;
 
-  public AttemptDetail(String channel, String attemptUuid, String status, String time) {
+  public AttemptDetail(String channel, String attemptUuid, String status, String time, String brandName, String appHash) {
     this.channel = channel;
     this.attemptUuid = attemptUuid;
     this.status = status;
     this.time = time;
+    this.brandName = brandName;
+    this.appHash = appHash;
   }
 
   public AttemptDetail(){}
@@ -26,5 +30,11 @@ public class AttemptDetail {
   }
   public String getTime(){
     return time;
+  }
+  public String getBrandName() {
+    return brandName;
+  }
+  public String getAppHash() {
+    return appHash;
   }
 }